Huge Program Rewrite
With the previous version of this program, I wrote myself into a corner as the GameController was created in the GamePlayScene, so I moved it to the main class. Now, we can choose the number of players, name those players, and choose a piece for them.

I also added a private constructor to all of my utility classes as this is good practice.