Players can now buy hotels, and the board displays it as well!
This is the fastest I have tackled a use case for this project! In all seriousness, the buy hotel box uses a class very similar to the buy house box, so I just needed to change a few the different parameters and columns in the class for this to work.